requirements.txt
coverage==3.7.1
Django==1.8.2
django-cors-headers==1.1.0
django-nose==1.4
djangorestframework==3.1.3
nose==1.3.7
PyJWT==1.3.0
wheel==0.24.0
